requirements.txt
kafka-python
requests
pyspark